requirements.txt
# AskTheDocs - Dependências Python
# Aplicação web para interação com documentação usando RAG

# Framework web
streamlit>=1.50.0

# Processamento de linguagem natural e LLM
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.1.0
langchain-groq>=0.1.0

# Embeddings e busca vetorial
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# APIs externas
requests>=2.31.0

# Gerenciamento de ambiente
python-dotenv>=1.0.0

# Processamento de dados
pandas>=2.0.0
numpy>=1.24.0

# Utilitários
pathlib2>=2.3.7
